+++ src/main/java/domain/intention/DrawCardIntetion.java
@@ -1,4 +1,4 @@
-package domain.answer;
+package domain.intention;
 
 import java.util.Arrays;
